Sensors list
Soil Moisture sensor
 Rain gauge sensor 
Wind direction and speed sensor 
An Accelerometer 
Air Temperature sensor 
Soil Temperature sensor 
Humidity sensor Gas sensor

Soil Moisture - the basic requirement of laying for a female locust is that the moisture content of the soil has to allow the locust to make a hole at a depth of approximately 5-15 cm. If the female cannot embed the whole egg - pod in the moist soil, she will not lay in that site.
Rainfall level - an equivalent of 15-20 mm of rain for 24-48 hours provides the most acceptable sites before laying.
Wind direction and Speed - migration of swarms is always downwind, so if the direction and speed of the wind is known, the destination of the locust can be predicted
Accelerometer - topography influences when and where breeding occurs because highlands often receive more rain than lowlands. Run-offs lead to suitable breeding conditions tens of kilometers downstream from where the rain fell
Air & soil Temperature -  air temperatures of less than 17 degrees discourage breeding and egg laying. Egg development is often dependent of soil temperature which should be less than 35 degrees celcius.
Humidity - egg and hopper development is dependent on humidity levels of a given area.
Gas - specific gas released during breeding to be detected